batCreateNovel/requirements.txt
openai
python-dotenv==0.21.0
functions-framework==3.3.0